Change `run` command to shell function
Shell alias was only receiving the first argument passed. Instead a
shell function `run()` is now set so any/all arguments can be saved
in a local variable and passed as one to the container.

This will work as originally intended to execute any code in the
container without entering it.